Move all plotting into a single module
And move this module out of psi/, which is otherwise concerned only
with the actual model
